Update README.md
Updating, editing

For the public facing documentation, it would be useful to clarify, which files/directories are intended for use/downloading by the public. 
Lets add notes to this Readme on the location of the latest formal and informal releases.

More details are needed for the 'credits' section'. Not sure what this is referring to.

Thanks !!
Lynn