Update mixs.yaml area_samp_size definition
Currently in v6.2 the definitions of area_samp_size and samp_size are both identical. The current definition is "The total amount or size (volume (ml), mass (g) or area (m2) ) of sample collected."
The proposal here is for area_samp_size to have a cleaner more differentiated definition. (My preference would be that that this was explicitly an area so could be say 144 cm^2 rather than a dimension 12 cm x 12 cm. Could then cope with non-quadrangles etc., but that is for another time))